Reorganised package structure and added random numbergenerator
Reorganised package structure to have the apocalypseEntity, human and zombie in separate files. Also moved the Zombie and Human constructors into the server package. This allows us to pass into it the whole server datastructure (instead of just an IExposedServerFunctions interface) allowing us to construct agents with extended functionality.
