+++ environments/apocalypseServer.go
@@ -1,11 +1,12 @@
 package apocalypseServer
 
 import (
+	"math/rand/v2"
 	"time"
 	extendedAgents "zombieApocalypeSOMAS/agent"
 	"zombieApocalypeSOMAS/physicsEngine"
 
-	//"github.com/kyroy/kdtree"
+	"github.com/MattSScott/basePlatformSOMAS/v2/pkg/agent"
 	"github.com/MattSScott/basePlatformSOMAS/v2/pkg/server"
 )
 
@@ -17,6 +18,7 @@ type IApocalypseServer interface {
 
 type ApocalypseServer struct {
 	*server.BaseServer[extendedAgents.IApocalypseEntity]
+	randNumGenerator *rand.Rand
 }
 
 func CreateApocalypseServer(numberZombies, numberHumans, iterations, turns int, maxDuration time.Duration, maxThreads int) *ApocalypseServer {
@@ -24,14 +26,38 @@ func CreateApocalypseServer(numberZombies, numberHumans, iterations, turns int, 
 		BaseServer: server.CreateServer[extendedAgents.IApocalypseEntity](iterations, turns, maxDuration, maxThreads),
 	}
 	for i := 0; i < numberZombies; i++ {
-		zombie := extendedAgents.SpawnNewZombie(10.0, physicsEngine.Vector2D{X: 0, Y: 0}, server)
+		zombie := server.SpawnNewZombie(10.0, physicsEngine.Vector2D{X: 0, Y: 0})
 		server.AddAgent(zombie)
 	}
 	for i := 0; i < numberHumans; i++ {
-		human := extendedAgents.SpawnNewHuman(10.0, physicsEngine.Vector2D{X: 0, Y: 0}, server)
+		human := server.SpawnNewHuman(10.0, physicsEngine.Vector2D{X: 0, Y: 0})
 		server.AddAgent(human)
 	}
 	return server
+}
+
+func (serv *ApocalypseServer) SpawnNewHuman(mass float32, initialPosition physicsEngine.Vector2D) *extendedAgents.Human {
+	entity := &extendedAgents.ApocalypseEntity{
+		BaseAgent:     agent.CreateBaseAgent(serv),
+		PhysicalState: physicsEngine.CreateInitialPhysicalState(&initialPosition, mass),
+	}
+
+	human := &extendedAgents.Human{ApocalypseEntity: entity}
+	return human
+}
+
+func (serv *ApocalypseServer) SpawnNewZombie(mass float32, initialPosition physicsEngine.Vector2D) *extendedAgents.Zombie {
+	entity := &extendedAgents.ApocalypseEntity{
+		BaseAgent:     agent.CreateBaseAgent(serv),
+		PhysicalState: physicsEngine.CreateInitialPhysicalState(&initialPosition, mass),
+	}
+
+	zombie := &extendedAgents.Zombie{
+		ApocalypseEntity: entity,
+		RandNumGenerator: serv.randNumGenerator,
+		Strength:         10,
+	}
+	return zombie
 }
 
 func (server *ApocalypseServer) GetNumEntity(entity extendedAgents.Species) int {
